+++ spec/lib/sequel_rails/railties/database_rake_spec.rb
@@ -1,9 +1,10 @@
 require 'spec_helper'
 
-describe 'Database rake tasks' do
+describe 'Database rake tasks', no_transaction: true do
 
   let(:app) { Combustion::Application }
   let(:app_root) { app.root }
+  let(:schema) { "#{app_root}/db/schema.rb" }
 
   around do |example|
     begin
@@ -15,8 +16,6 @@ describe 'Database rake tasks' do
   end
 
   describe 'db:schema:dump' do
-    let(:schema) { "#{app_root}/db/schema.rb" }
-
     it "dumps the schema in 'db/schema.rb'" do
       Dir.chdir app_root do
         `rake db:schema:dump`
@@ -59,6 +58,33 @@ EOS
           :schema_migrations
         ).insert_sql(:filename => '1273253849_add_twitter_handle_to_users.rb')
         expect(File.read(schema)).to include sql
+      end
+    end
+  end
+
+  describe 'db:rollback' do
+    it 'revert latest migration' do
+      Dir.chdir app_root do
+        begin
+          expect do
+            `rake db:rollback`
+          end.to change{ SequelRails::Migrations.current_migration }.from(
+            "1273253849_add_twitter_handle_to_users.rb"
+          ).to(nil)
+        ensure
+          SequelRails::Migrations.migrate_up!
+        end
+      end
+    end
+  end
+
+  describe 'db:migrate:redo' do
+    it 'run down then up of the latest migration' do
+      Dir.chdir app_root do
+        SequelRails::Migrations.migrate_up!
+        expect do
+          `rake db:migrate:redo`
+        end.not_to change{ SequelRails::Migrations.current_migration }
       end
     end
   end