+++ spec/spec_helper.rb
@@ -22,8 +22,12 @@ RSpec.configure do |config|
   config.treat_symbols_as_metadata_keys_with_true_values = true
   config.filter_run_excluding rspec_exclusions
   config.around :each do |example|
-    Sequel::Model.db.transaction(:rollback => :always) do
+    if example.metadata[:no_transaction]
       example.run
+    else
+      Sequel::Model.db.transaction(:rollback => :always) do
+        example.run
+      end
     end
   end